backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
PyGithub==2.1.1
python-dotenv==1.0.0
pydantic==2.5.3
aiosqlite==0.19.0
python-multipart==0.0.6 